VCarveSettings and BitmapSettings popup window class
